[code-merger][code-gen] fix/improve handling slot code merging
We now emit a JSXFragment around defaultNodes to accurately detect the
boundaries between defaultNodes. This fixed a bug where single default
node will capture the "args.<argName>" as the chrome of the defaut node,
which causes merging to be wrong.

For historical code, we also change the boundary detection algorithm to treat
the JSXElement as the boundary. This will cause previous code to emit
the rh.showXXX function twice after code-merger. The work around is to
just add a fragment around the default node.

Change-Id: I50b0a76ec204eb03aee7aae119aab7d799ae85b4