[cli] cleanup and add append-jsx-on-missing-base for migration
Now that we generate stable id, we don't need to parse the skeleton file
before codegen. We also don't need to dump metadata to mergeDumpDir
given that now we dump to the helper file.

Add append-jsx-on-missing-base option so that when we perform update for
code generated before the stable sequential id was introduced, developer
can perform a manual merge.

Also upgrade version of cli, code-merger.

Change-Id: I7dd5cde148304bddf1089efd2bd823af3b7255d6