initial CSS modules support
Changes not ready to release yet, we want to improve the class names when using CSS modules. It should work for the JS version of Plasmic Levels, for both blackbox and direct edit settings.

There're several "If" checks in this change, adding to the existing logic. This is intentional. I initially wanted to abstract the way we create "class names", but I think a much bigger abstraction is required. One that encompasses all of the "styles" changes, together with a new "component.addProperty" API.

Change-Id: If1cf7a086e5aff296159dd5d444dc68dd4f6fdff