Add support to Next.js and Gatsby <Link> component
There are several possible approaches to add support for Next.js and
Gatsby <Link> components. This patch does that by adding a PlasmicLink
component to react-web, which is a platform-aware link wrapper.

The pro about this approach is that it is not necessary to add much
complexity to codegen (e.g. no new nodes on tplTree), and splitting
`props` between <Link> and <a> (in Next.js case) is straightforward.
User can override props like it's only one element and react-web
will split the props between them, e.g.

  <PlasmicHomePage
    link={{
      className: "class-to-a-element", // this will go to <a>
      href: "/a-nice-link",            // this will go to <Link>
      replace: true,                   // this will go to <Link>
    }}
  />

The con is requiring a wrapper instead of simply using platform's
<Link>. I think that's fine because for Gatsby we would need a
wrapper anyway to decide if the links are internal or external
(https://www.gatsbyjs.com/docs/reference/built-in-components/gatsby-link/#reminder-use-link-only-for-internal-links).

To avoid requiring gatsby-link and next-link in react-web, PlasmicLink
receives the platform component in a "component" prop. So the generated
code (for blackbox) looks like the following:

  <PlasmicLink
    data-plasmic-name={"link"}
    data-plasmic-override={overrides.link}
    className={classNames(
      sty.link,
      defaultcss.all,
      defaultcss.__wab_text
    )}
    component={Link}
    href={"/test-page"}
    platform={"gatsby"}
  >
    {"Link children."}
  </PlasmicLink>

Change-Id: Ibe7e4232fdf7456be2cfae3324250528fb38e9e3