[bug/cli] Fix projectCss checksum and run plasmicLock migrations
The projectCss checksums were not being stored (but rather the
projectCss string). This is to fix the bug and extends the migration
framework to also update existing plasmic.lock files.

It also ensures that projectLock.fileLocks are always set, simplifying
the code around it.

Issue: https://app.clubhouse.io/plasmic/story/9039/projectcss-checksums-are-not-being-stored-correctly
Change-Id: I0da651d701caf2066cc00afba68fd4274088dc7f