Add per-project tokens
https://app.clubhouse.io/plasmic/story/10408/move-away-from-secret-auth-tokens

No longer require a user API token to sync.

Per-project tokens are now returned from the server and persisted into a
plasmic.json file. It is safe to include these directly in your source repo or
make them publicly accessible.

As a result:

- You can perform Plasmic operations on CI/CD without needing ta add Plasmic
  secret management to to your jobs.
- You no longer need everyone on a team to authenticate with Plasmic before
  they can sync.
- Users no longer have to think about whether they should be creating an "ops"
  account or using a designated user's account for automated jobs.
- Users have to think about whether and how to create and manage multiple
  Plasmic auth tokens.

You can still continue to use a user API token (this is how you are allowed to
get the per-project tokens in the first place).

This comes with server changes.

Still to-do: support this for PlasmicLoader.

Change-Id: Ic228ee6b088dd70a5235713c7415a97504832891