[token] Add CLI support for plasmic-loader.json
https://app.clubhouse.io/plasmic/story/10408/move-away-from-secret-auth-tokens

And bump version.

We introduce a plasmic-loader.json file that holds the per-project API
tokens for each of the root projects.

Over time we may want to nudge users to put more of their config in
plasmic-loader.json rather than editing next/gatsby configs, since we
can build tooling around the former.

We also introduce a --loader mode flag to the CLI, which PlasmicLoader
will use.

Change-Id: I698a5597077a6d669ac1378cbf681698d3bf6a58