[cli] Fix version of project dep saved to config on first sync
`syncProject` was running `syncCodeComponentsMeta` before
`syncProjectConfig` causing all project dependencies to be saved with
version "latest" in plasmic.json (because the function that syncs code
component meta runs `getOrAddProjectConfig` with an undefined base).

That way, when an user was first syncing a project X that depends on
Y@1.0.0 the CLI was syncing Y@1.0.0 but saving Y@latest do plasmic.json.

This patch fixes that behavior by saving Y@1.0.0 as expected.

Change-Id: I7fed6df7768c7bfb22c7196f3849ed0d4386ab5d