[host] Fix re-importing host and create a module for `registerComponent`
Importing `registerComponent` directly from `@plasmicapp/host` leads to
issues if we import it more than once in the same window, since `__Sub`
will be re-recrated and overwrite the existing value.

This is to avoid re-creating __Sub if it already exists.

Also if we don't expect to actually host the components when registering them
(e.g., when defining host-less code components) we should import
directly from `@plasmicapp/host/registerComponent` to reduce bundle size

Change-Id: I47c8f9f04829eefce0bceed0ff42719b4a2808b8