[loader] Fix revalidate for Next 12
It seems that in Next 12, in a static regeneration call, Next is
now re-using processes?  And so the same instance of
PlasmicComponentLoader may now be reused across different calls
to regenerate.  That means we need to make sure we by-pass the
in-memory cache so that regenerate will fetch the latest data.

Unfortunately this will probably slow down the `yarn build` phase
because now we need to do a fetch for every page, but I don't think
it's possible to tell if we're in `yarn build` or in regenerate.

Change-Id: I219ea700a013b53495f949f435a0232e2feaa451