[host] Add "array" type, "isRepeated" param and "defaultExpr"
When `array` is used, studio will render a "custom code expression"
field, a JavaScript editor that produces code to be evaluated in canvas
with the data bindings. It's different from the existing `code` type
because that returns a string and supports multiple languages. The
idea is that at the moment studio will force fields of `array` type to
be permanently "switched to a custom code expression". In the future
we may have a different UI for arrays.

`isRepeated` will be used to allow us to give a special treatment for
repeater components in studio.

`defaultExpr` is used to allow specifying a default expression instead
of a default value.

This is in preparation for the Repeater component (that will be
introduced in plasmic-basic-components once `host` is published with
the changes in this patch).

Change-Id: I2b44f4c862bd818a4bb597e89b0864a0df5afaa2