[host] Add "hidden" prop to DataProvider
This patch adds a `hidden` prop to DataProvider. That prop can be used
to hide something in data picker and custom code expression editor in
Plasmic studio.

A new type DataMeta is used to specify meta data for data context
fields. Right now this type has only `hidden`, but it can have other
things in the future (e.g. display name and description).

When we do:

  <DataProvider name="foo" value="bar" hidden={true} />

... that will now add two entries to data context instead of one. We
will get the following context:

  $ctx = {
    foo: "bar",
    __plasmic_meta_foo: {
      hidden: true
    }
  }

Note that keys starting with __plasmic_meta will always be hidden in
data binding.

Change-Id: I946e2f6a401d2b9c02d8b6556d34656f57cc811d