[loader] Don't use a singleton registry in loader
This was introduced in https://gerrit.aws.plasmic.app/c/public-packages/+/8096/13/packages/loader-react/src/loader.ts#104

But, the server/browser modules are cached on the loader, not in the
registry, so it doesn't seem necessary to have a singleton registry?

However, it is harmful, because when we do registry.registerModules(),
we are writing to the singleton registry; this is bad for
"@plasmicapp/loader-runtime-registry", which points to
loader.substitutedComponents on the loader instance. That means
if you instantiate multiple loaders, @plasmicapp/loader-runtime-registry
will be pointing to the last one's loader.substitutedComponents, which
may be bad if the last one isn't actually used to register components
(and so is empty :-/ )

To get around loader-gatsby using different instances of loader at
load and render, had to also move the fetched bundle caching out to
loader-fetcher :-/

Change-Id: I2499650e5f02e60c3f93e1ddd5568d575378e77a