[state-management] add overrideEnv to initFunc
Initializing states from code components can be a bit tricky.
Let's say we have a code component called "Counter" with an exposed state named "count".
We need to initialize this state variable in the parent component. Typically, we would use "$props" to initialize the state variable, but in this case, "$props" refers to the parent component instead of the child component that we need to use.

Thus, we need a method to provide an alternative "$props" to the code components initFunc.

Change-Id: I0f8ef3af9994e0d009ffde2cac712928ccffde95