[react-web] Add hack for exposing PlasmicImg in generated code
If you use codegen, and you use `declaration: true` in your
tsconfig, and you have an image, you will see this error on this
line of code:

  export const PlasmicX = Object.assign(
    makeNodeComponent("root"),
    {
      img: makeNodeComponent("img"),
      ...
    }
  );

The problem is exporting the return value of `makeNodeComponent("img")`
which is a component that takes in props of PlasmicImg.  Typescript
will complain:

  The inferred type of 'PlasmicX' cannot be named without a reference to '../../../node_modules/@plasmicapp/react-web/dist/render/PlasmicImg'. This is likely not portable. A type annotation is necessary.

But actually the type of PlasmicImg is properly exported from
@plasmicapp/react-web!  This is actually caused by an open
typescript issue, and this CR is implementing a hacky workaround
mentioned here: https://github.com/microsoft/TypeScript/issues/47663#issuecomment-1595614663

So, for a package that uses "exports" (which we do), we need to add
some fake "export path" that explicitly names the file that is
needed (even though that file is already reachable and referenced
from `./dist/index.d.ts`).

Sigh.

Change-Id: I98090db5a5eb0d8dacee9189b9bb5580f8343fec