[wab] Fix custom arenas being loaded out of view
For custom arenas only, if the page was loaded and the auto-zoom doesn't need to change the zoom (zoom = 1), then .canvas-editor__scaler doesn't have a transform style, which means it will be absolutely positioned by default instead. This resulted in the frame being rendered at (0,0), but the clipper is centered, and so the frame is out of view.

To fix this, we can simply add a default transform: scale(1) to .canvas-editor__scaler.

Issue: https://linear.app/plasmic/issue/PLA-10348
Change-Id: I8d30d4c58a9872c9f6e5e1728aa16f535c697e85
GitOrigin-RevId: 2ec5a858eb21cc675dc6a4db09b1e63511c9e983