[cli] allow files to not be overwritten
This is useful when a Plasmic* file is deleted and marked to recreation. When syncing the component again, all three files (PlasmicComp, Comp and Comp.module.css) are synced again. In this case, we want to be able to answer NO to the override of `Comp` because it can have user code

Issue: https://linear.app/plasmic/issue/PLA-10285/dont-prompt-when-overwriting-plasmic-files-during-plasmic-sync
Change-Id: Ia9665440686ac193097a7923aa9ccb9d56bfd8ba
GitOrigin-RevId: f37dfa10b3fccd1215b9336a02ab73eedbfd8292