Fix token resolution for displaying registered tokens
Previously, the resolve functions only return the resolved value,
but this doesn't work for registered tokens with selector,
since the selector data is lost.

This commit fixes it by introducing a new set of functions
that also returns the StyleToken. With the StyleToken,
the selector element can then be queried to find the token.

Issue: https://linear.app/plasmic/issue/PLA-10750
Change-Id: I7ce0b6e840cc109582f767871c953e5ef8e48c25
GitOrigin-RevId: 1d066da208e86d968ada67d3b22c15c60cad3786